+++ contracts/bondingcurve/IBondingCurve.sol
@@ -19,18 +19,18 @@ interface IBondingCurve {
 	/// @param to address to receive FEI
 	/// @param amountIn amount of underlying tokens input
 	/// @return amountOut amount of FEI received
-	function purchase(address to, uint256 amountIn) external payable returns (uint256 amountOut);
+	function purchase(address to, uint amountIn) external payable returns (uint amountOut);
 	
 	// ----------- Governor only state changing api -----------
 
 	/// @notice sets the bonding curve price buffer
-	function setBuffer(uint256 _buffer) external;
+	function setBuffer(uint _buffer) external;
 
 	/// @notice sets the bonding curve Scale target
-	function setScale(uint256 _scale) external;
+	function setScale(uint _scale) external;
 
 	/// @notice sets the allocation of incoming PCV
-	function setAllocation(address[] calldata pcvDeposits, uint256[] calldata ratios) external;
+	function setAllocation(address[] calldata pcvDeposits, uint[] calldata ratios) external;
 
 	// ----------- Getters -----------
 
@@ -41,24 +41,24 @@ interface IBondingCurve {
 	/// @notice return the average price of a transaction along bonding curve
 	/// @param amountIn the amount of underlying used to purchase
 	/// @return price reported as FEI per X with X being the underlying asset
-	function getAveragePrice(uint256 amountIn) external view returns (Decimal.D256 memory);
+	function getAveragePrice(uint amountIn) external view returns (Decimal.D256 memory);
 
 	/// @notice return amount of FEI received after a bonding curve purchase
 	/// @param amountIn the amount of underlying used to purchase
 	/// @return amountOut the amount of FEI received
-	function getAmountOut(uint256 amountIn) external view returns (uint256 amountOut); 
+	function getAmountOut(uint amountIn) external view returns (uint amountOut); 
 
 	/// @notice the Scale target at which bonding curve price fixes
-	function scale() external view returns (uint256);
+	function scale() external view returns (uint);
 
 	/// @notice a boolean signalling whether Scale has been reached
 	function atScale() external view returns (bool);
 
 	/// @notice the buffer applied on top of the peg purchase price once at Scale
-	function buffer() external view returns(uint256);
+	function buffer() external view returns(uint);
 
 	/// @notice the total amount of FEI purchased on bonding curve. FEI_b from the whitepaper
-	function totalPurchased() external view returns(uint256);
+	function totalPurchased() external view returns(uint);
 
 }
