+++ contracts/pcv/UniswapPCVDeposit.sol
@@ -16,26 +16,26 @@ abstract contract UniswapPCVDeposit is IPCVDeposit, UniRef {
 	constructor (address core, address _pair, address _router, address _oracle) public
 		UniRef(core, _pair, _router, _oracle) {}
 
-	function withdraw(address to, uint256 amountUnderlying) external override onlyPCVController {
-    	uint256 totalUnderlying = totalValue();
+	function withdraw(address to, uint amountUnderlying) external override onlyPCVController {
+    	uint totalUnderlying = totalValue();
     	require(amountUnderlying <= totalUnderlying, "UniswapPCVDeposit: Insufficient underlying");
 
-    	uint256 totalLiquidity = liquidityOwned();
+    	uint totalLiquidity = liquidityOwned();
     	Decimal.D256 memory ratioToWithdraw = Decimal.ratio(amountUnderlying, totalUnderlying);
-    	uint256 liquidityToWithdraw = ratioToWithdraw.mul(totalLiquidity).asUint256();
+    	uint liquidityToWithdraw = ratioToWithdraw.mul(totalLiquidity).asUint256();
 
-    	uint256 amountWithdrawn = _removeLiquidity(liquidityToWithdraw);
+    	uint amountWithdrawn = _removeLiquidity(liquidityToWithdraw);
     	_transferWithdrawn(to, amountWithdrawn);
     	_burnFeiHeld();
     	emit Withdrawal(msg.sender, to, amountWithdrawn);
     }
 
-	function totalValue() public view override returns(uint256) {
-		(, uint256 tokenReserves) = getReserves();
+	function totalValue() public view override returns(uint) {
+		(, uint tokenReserves) = getReserves();
     	return ratioOwned().mul(tokenReserves).asUint256();
     }
 
-	function _getAmountFeiToDeposit(uint256 amountToken) internal view returns (uint amountFei) {
+	function _getAmountFeiToDeposit(uint amountToken) internal view returns (uint amountFei) {
 		(uint feiReserves, uint tokenReserves) = getReserves();
 		if (feiReserves == 0 || tokenReserves == 0) {
 			return peg().mul(amountToken).asUint256();
@@ -43,8 +43,8 @@ abstract contract UniswapPCVDeposit is IPCVDeposit, UniRef {
 		return UniswapV2Library.quote(amountToken, tokenReserves, feiReserves);
 	}
 
-    function _removeLiquidity(uint256 amount) internal virtual returns(uint256);
+    function _removeLiquidity(uint amount) internal virtual returns(uint);
 
-    function _transferWithdrawn(address to, uint256 amount) internal virtual;
+    function _transferWithdrawn(address to, uint amount) internal virtual;
 
 }
